[networkimage] Use the passed-in content mode when checking the in-memory cache for the image.
This was causing a bug where images would thrash the in-memory cache
when a content mode was passed into the method that was different
from the content mode of the view.

This makes sure that we're checking the same content mode that we
end up passing to the thread which we then end up passing back to
the image memory cache to store the image, completing the full
circle of images in memory.

Tests: Create a network image view. Set the content mode to
UIViewContentModeCenter. Load a network image with any other
content mode. Load the image again and verify that it loads from
memory.