[photos] Fix misc bugs related to animating between photos and orientation changes.
Most of these bugs were found when the device changed orientation during
an animation to another image. This would cause the photo album view to
go crazy as it tried to simultaneously update the page layouts while
animating. The solution was to simply ensure that we maintain the
state of whether we're programmatically modifying the content offset
during a rotation to avoid updating page layouts during this time.