[photos] Use the original photo dimensions to scale the thumbnail to the final image size to avoid the "snapping" effect.
This is a pretty awesome extra feature for the Nimbus photo library
that is a good example of something non-trivial made easy with Nimbus.
If we know the final image's dimensions from the API then we can inform
Nimbus of these dimensions so that it can size the thumbnail to match
the final scale of the image. This allows us to perfectly replace the
thumbnail with the final image once it has loaded without any visual
snapping.