[photos] Refactor the code related to zoomingAboveOriginalSizeIsEnabled and fix some bugs with thumbnails snapping to the original photos on load.
Added zoomingAboveOriginalSizeIsEnabled to the NIPhotoAlbumScrollView so that
it is configurable from the controllers.

The way that the thumbnail min zoom size was calculated before when we knew
the final photo dimensions was complicated and not correct. This makes the
logic much simpler. If the final image will be smaller than the view, we
calculate how much we need to scale the thumbnail by such that it perfectly
fits where the final image will be by multiplying the scale against the
relative size of the two images.