[webcontroller] Fix a memory leak caused by not releasing views on dealloc.
This is entirely due to an artifact of Three20. In Three20, every
controller that subclasses TTViewController will have viewDidUnload
called on dealloc, so it wasn't necessary to release subviews in
both places. This is not the case in Nimbus whatsoever, so we must
be more conscious about releasing subviews explicitly in both the
dealloc and viewDidUnload.