[overview] Add the overiew bundle to the Nimbus project. Fix bug with the overview when not using navigation controllers.
When not using a navigation controller the overview would not correctly
intercept the status bar height. This would result in the root view controller
taking up the entire screen's real estate rather than leaving a bit of
room for the overview. The fix simply involves swizzling one more method
on UIApplication for fetching the status bar height and returning the Overview
height when applicable.