[core] Don't retain the node in the NILinkedListLocation object.
This resolves a retain cycle that was occurring in the NIMemoryCache
object. The retain cycle was occurring because the lruLocation was
being retained by the NIMemoryCacheInfo object which was itself
being retained by the NILinkedListLocation object's node property.
Making the node property non-retaining removes this retain cycle.

Unit tests all pass and no memory leaks were detected in the photos
app when pushing and popping view controllers.