[photos] Fix bugs in the layout of the photo scroll views.
Due to pushing the scroll view down to a subview I needed to update
the way we calculate centering the photo. This needed to be done
when the scrollview's layoutSubviews method was called so I created
a simple subclass of UIScrollView that handles this. The rest of
the photo scroll view functionality was able to stay in the root
of the view.

Also fixed some issues with photo centering. The problem was related
to when we set the frame of the paging scroll view and the fact that
it was simply setting the frame of all of the subviews, without giving
them the opportunity to maintain their existing state.