Add namespacing support to Nimbus.
Namespacing is meant for parties that wish to distribute a static library that
uses any part of Nimbus without causing duplicate symbol errors if someone
uses both the static library and Nimbus. This has the downside of potentially
duplicating some code, but the benefit of guaranteed Nimbus functionality within
the static library outweighs this downside.

To use the Nimbus namespace, import the +Namespace header for each feature you use
immediately before the feature's header and define __NIMBUS_NAMESPACE_PREFIX_ to
whatever you want Nimbus' prefix to be.