Set the text field tag and adjust font size to fit width flag.
I set the text field tag to the element ID for identifying the text
field when using the delegate methods. This is particularly useful when
you're attempting to validate the data within the text field and you
need to know which one it is. Setting the adjustFontSizeToFitWidth flag
also removes the issue of truncated placeholder or text field text.