Forward all UIScrollViewDelegate methods to the NIPagingScrollView delegate
NIPagingScrollView claims to implement the UIScrollViewDelegate protocol, but will silently swallow most of these methods.

This is annoying in many cases. For instance, it prevents the clients of NIPagingScrollView to implement properly an UIPageControl with deferred updates (as deferring the updates requires the – scrollViewDidEndScrollingAnimation: method).

Instead of manually implementing and forwarding all the methods UIScrollViewDelegate methods methods, this patch uses the NSObject message forwarding facility. It's much shorter than the manual implementation, not much more complicated, and safe if new methods are UIScrollViewDelegate in the future.