[attributedlabel] Links now have a gutter area.
Fingers are clumsy. Sometimes, when you tap a link, your finger will
move subtly right as you're lifting your finger, causing the attributed
label to think that you moved away from the link and that no tap should
be registered.

A "gutter" introduces a 22px padding on all sides of the tapped link.
Touches within this gutter will still register as touches within the
link, even if your finger is actually physically within a different link.

This "feels" right because we highlight the originally-tapped link to
let the user know which link is active.

This addresses some potential issues related to issue #236.