Ensure photo zooming is set correctly after page reuse.
NIPhotoScrollView's setZoomingIsEnabled: checks whether an image is set
before allowing the zoomingIsEnabled property to affect its internal
scroll view's zoom.

This is usually fine whilst scrolling to the left of the album to get
the next pages, however when scrolling back and loading previous pages
whose images have been cleared due to prepareForReuse, the call to
setZoomingIsEnabled: is too early to have the intended effect.

This patch remedies this by setting the property only after any
potential image is set on the page.