Cancel imageProcessingBlock on new network request
Problem: setPathToNetworkImage:forDisplaySize:contentMode:cropRect:
cancels any currently running network operations. But this is not 
enough if the networking part of the operation has already finished but
the associated imageProcessingBlock is still executing.

By storing the most recently requested network path at the beginning of
the method and later comparing it to the network path that is associated
with the imageProcessingBlock, we only use the result of the block if
the two are identical.

See https://github.com/jverkoey/nimbus/issues/389 for reference.