NIPagingScrollView additions
NIPagingScrollView.pageInset if non zero will inset the page from full screen, thus making neighboring pages visible.  This is useful for showing that there is scrollable content and presents a 'film-strip' like appearance.

Note in this case we take on the page snapping responsibilities as UIScrollView is fundamentally view bounds oriented and doens't support edges of other pages showing.