Fallback to regular usage of UIAccessibilityElement when the accessibility container is changed externally.
NIViewAccessibilityElement overrides UIAccessibilityElement behavior by
delaying computing its accessibility frame (which is in screen corrdinates).
In that case, the accessibility frame is computed upon request based on the
position at that moment of the actual visual component made accessible.

This holds true until the accessibility container is changed externally. For
example, when the accessibility element is in a table view hierarchy). In that
case, the initial accessibility container is changed to a
UITableViewCellAccessibilityElement and frameInContainer becomes invalid.

This commit invalidates NIViewAccessibilityElement's special treatment and
fallback to the regular behavior.