[attributed-label] Add delegate method for long presses on attributed label data detector results 👇
This change introduces a new delegate method,
-attributedLabel:didLongPressTextCheckingResult:atPoint:, which is used
to detect when a user long presses on a data detector result. This
delegate method enables projects to avoid the creation of UIActionSheet
objects (the UIActionSheet class is deprecated) and avoid implementing
-attributedLabel:shouldPresentActionSheet:withTextCheckingResult:atPoint:
which contains the deprecated type in its signature. Clients can
implement -attributedLabel:didLongPressTextCheckingResult:atPoint: and
present a UIAlertController appropriately. It is not appropriate for
NIAttributedLabel to attempt to determine an appropriate position in the
view controller hierarchy from which to present a UIAlertController.
The user interaction **must** be delegated to an object that has
appropriate access to the view controller hierarchy to properly present
a UIAlertController. As this new delegate method represents the
migration path away from legacy UIActionSheet presentation, responding
to this selector is treated as opting out of action sheet presentation
and overrides
-attributedLabel:shouldPresentActionSheet:withTextCheckingResult:atPoint:.

This change also introduces a testing header and some tests to verify
the desired delegate callback behaviors.

Tested by running tests with iPhone XR simulator running iOS 12.1.