Fix compilation with -Wextra-semi
The clang flag -Wextra-semi warns about semicolons that are in
a place incompatible with the c++98 standard. As clang applies
the flags used for source file to all the included header files,
it is necessary to fix header files to enable the flag in
third-party projects.

This changelist removes the extra semicolons in header files
in order to allow inclusion of the header from file compiled
with -Wextra-semi flag enabled.

This fixes the following error (path omitted):

  In file included from ...
  .../NIDeviceOrientation.h:88:2: error: extra ';' outside a function is incompatible with C++98 [-Werror,-Wc++98-compat-extra-semi]
  };