Fix for wrong accessibilityFrames in NIAttributedLabel (#621) (#723)
UIKit sometimes changes the accessibility containers of all accessibility
elements in a NIAttributedLabel to another view. I have observed this happening
when an NIAttributedLabel is embedded within a UICollectionViewCell, in which
case the newly assigned accessibility container is always the cell itself. I am
unsure why this happens.

We do compute a fallback value for accessibilityFrame for each
NIViewAccessibilityElement we create, but they become invalid as soon as the
label's position on screen changes. This frequently happens when the label is
embedded within a scroll view like UICollectionView.

When an NIViewAccessibilityElement has an invalid accessibilityFrame, its
activation point also becomes invalid, causing it to be untappable by VoiceOver
users.

This PR allows NIViewAccessibilityElement to remember its last valid
accessibility container upon request. When computing a dynamic accessibility
property like accessibleFrame, it will attempt to use the last valid
accessibility container it remembers if there is one instead of falling back to
super.